pyproject.toml
[project]
name = "make-sps"
version = "1.0.0"
description = "make sps file"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "coverage>=7.8.2",
    "fastapi>=0.115.12",
    "flake8>=7.2.0",
    "pillow>=11.2.1",
    "pydantic>=2.11.5",
    "pyflakes>=3.3.2",
    "pylint>=3.3.7",
    "pytest>=8.4.1",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "uvicorn>=0.34.2",
]

[ollama]
model = "devstral:24b"
apiBase = "http://localhost:11434"
